escape_quotes: Correct escaping in the presence of nested quotes
In most cases, nested quotes are not needed. When dealing with expressions
passed to find, they can appear, though. In these cases it becomes relevant to
escape backslashes that may already be present in the input string in addition
to handling the regular string escapes.

This commit changes the escape_quotes function accordingly.