Add auxiliary functions for building find() expressions
Previously, queries passed to find() had to be provided as correctly-quoted
strings directly to the function. Since this way, not much of the abstraction
features of such a library are leveraged, this commit replaces the previously
created string-based API with one that takes a tuple that represents
FILTER expression constructed of Erlang terms.

Usage could e.g. look as follows:

	erlmpd:find(Conn, {and, [{lnot, {tagop, artist, eq, "val"}},
	                         {tagop, title,  eq, "val2"}]}).