Revise sticker handling
Previously, unprocessed data was returned by most of the sticker functions.
This would require every application to handle this part of the MPD protocol
for itself.

This commit revises the sticker API to become more user-friendly:

 * It splits out all of the different functions (find/delete/get/...) to
   dedicated functions. This makes documenting the individual routines
   easier and also allows clear type specifications to be formulated.

 * It adjusts all the “reading” functions to parse the sticker values and
   discard redundant info (such as but not limited to the leading "sticker: "
   prefix etc). As a result, applications can now directly process the values
   they are interested in.

Note that currently, the find function is the only one to return binaries
wherehas the other ones work on strings. It is not exactly clear why this
happens, though. It may turn out that in the end the return value is better
processed as iolist?