idle: Pass errors returned from command
Previously, errors returned by the `idle` command were ignored.
Occasionally, when interacting with `noidle` and `idle`, errors of type
`ealready` were observed.

The correct solution in these cases is to ensure a proper alternation of
`idle` and `noidle` commands on the TCP level. This requires a means for idle()
to report when it has completed the `gen_tcp:send` of the CommandStr in order
for an application to only call `command("noidle")` after having received this
confirmation. This is complicated to implement on the library side and not much
easier on the application side.

As a workaround for the issue, this commit changes the `idle()` function to
return the error. It was checked that applications can handle the error and
continue to use the connection afterwards. A proper solution for this issue
could be implemented by a later commit.

The error handling provided by this commit should be fine independently of
future changes to the idle handling to fix the described `noidle`interaction.