count/2, count_group/3: Add new count APIs
Previously, an interface to MPD's count API was only provided by count/3 with
the intention of querying the count for exactly one tag match. It could already
be used for very simple “group” queries by telling the API that the tag of
interest was `group`. However, this resulted in the output list being difficult
to parse for the application and filtering could not be used in conjunction with
this API.

This commit attempts to solve these limitations as follows:

 * It adds a new count/2 API that is intended to produce an output format equal
   to the count/3 call but has enhanced querying capabilities by allowing
   arbitrary filter expressions to be used in place of a single tag match.

 * It adds a new count_group API that can be used to query multiple counts at
   once by grouping after a certain tag. It parses the resulting output inside
   the erlmpd library and outputs a list of property lists similar to other
   (e.g. song-based) APIs.