Partitions: Expose partitions features through new API calls
Previously, partition management was not explicitly exposed through erlmpd.
While it could be invoked through the generic `command` API this moved the
burden of properly parsing the protocol results to the client.

This commit adds new dedicated API calls for partition management. The names
are equal to the ones used in the MPD protocol specification.