sticker_find/7: Provide API for sticker searching with filter
Previously, sticker searching was only possible by name.
This commit adds support for the MPD protocol:

	sticker find {TPYE} {URI} {NAME} = {VALUE} [sort {SORTTYPE}] [window {START:END}]

The new functionality is provided by sticker_find/7. Although it is the first
function to allow making use of MPD's output sorting and pagination
capabilities, these have been implemented in a manner that allows extending the
same mechanism to other commands as needed, too.

The idea is to have an `Options` last argument that may take a proplist with
optional arguments including those for `window` and `sort`. Note that the
valid values to pass for `sort` differ between the standard `find` and
`sticker find`.