albumart, readpicture: Add new commands
Previously, erlmpd didn't support parsing binary responses from MPD. These kinds
of responses are defined to be returned by fucntions related to reading image
data such as the `albumart` and `readpicture` APIs.

This commit adds support for commands which may return binary data and makes use
of this to implement albumart/2 and readpicture/2. Unlike the commands on the
MPD level, these commands block until all of the data was returned. This makes
application-side processing easier at the expense of losing some flexibility
wrt. downloadnig image contents in the background.

The suggested implementation for users that need to do this in the background is
to offload the downloading of the image data to a separate process and
connection.